+++ Src/RiderPlugin/src/main/kotlin/com/jetbrains/rider/roflcopter/settings/MismatchedFileNamesOptionsPage.kt
@@ -3,8 +3,9 @@ package com.jetbrains.rider.roflcopter.settings
 import com.jetbrains.rider.settings.simple.SimpleOptionsPage
 
 class MismatchedFileNamesOptionsPage : SimpleOptionsPage(
-        "Mismatched file names",
-        MismatchedFileNamesOptionsPage::class.simpleName!!) {
+        name = "Mismatched file names",
+        pageId = MismatchedFileNamesOptionsPage::class.simpleName!!
+) {
 
     override fun getId(): String {
         return "preferences." + this.pageId;