+++ Src/Roflcopter.Plugin.Tests/MismatchedFileNames/MismatchedFileNameHighlightingQuickFixTest.cs
@@ -11,7 +11,7 @@ using Roflcopter.Plugin.MismatchedFileNames;
 namespace Roflcopter.Plugin.Tests.MismatchedFileNames
 {
     [TestFixture]
-    [TestNetFramework46]
+    [TestNet60]
     public class MismatchedFileNameHighlightingQuickFixTest : CSharpQuickFixTestBase<MismatchedFileNameHighlightingQuickFixTest.TestQuickFix>
     {
         private TestQuickFix _quickFix;