+++ History.md
@@ -1,3 +1,6 @@
+### 1.20.0 ###
+- ReSharper and Rider 2023.3 support
+
 ### 1.19.0 ###
 - ReSharper and Rider 2023.2 support
 - Add missing 'displayName' for options page in Rider plugin metadata